Dir(#80,#53) : 로그인, 회원가입 type - src/types/loginJoin 폴더로 이동, 소셜로그인 기능 폴더 및 파일 생성
