+++ src/types/loginJoin/types.ts
@@ -1,0 +1,11 @@
+export type SignUp = {
+  userId?: string;
+  email: string;
+  password?: string;
+  nickname?: string;
+};
+
+export type JoinApi = {
+  email: string;
+  password: string;
+};
